refactor(utils/generate => qri/generate): refactor based on pr comments
