+++ lib/components/editor/EditBody.js
@@ -9,7 +9,7 @@ import {
   generateSchemaFromRow,
   generateCollapsedChanges,
   generateMatchingSchemaAndBody
-} from '../../utils/generate'
+} from '../../qri/generate'
 
 import Base from '../Base'
 import ValidSelect from '../form/ValidSelect'
