+++ lib/components/editor/Editor.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { Route, Switch } from 'react-router'
 import { isEmpty, isEmptyObj } from '../../utils/reflect'
 
-import { generateParsedBodyAndStructure } from '../../utils/generate'
+import { generateParsedBodyAndStructure } from '../../qri/generate'
 
 import Results from './Results'
 import EditMeta from './EditMeta'
