+++ lib/qri/generate.js
