+++ lib/reducers/editor.js
@@ -20,7 +20,7 @@ import {
 
 import {
   generateMatchingSchemaAndBody
-} from '../utils/generate'
+} from '../qri/generate'
 
 import cloneDeep from 'clone-deep'
